requirements.txt
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
Babel==2.15.0
Brotli==1.1.0
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
cssselect2==0.7.0
defusedxml==0.7.1
distlib==0.3.8
Django==5.0.1
django-allauth==0.60.1
django-crispy-forms==2.1
django-crum==0.7.9
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
dnspython==2.5.0
filelock==3.13.1
fonttools==4.51.0
gunicorn==23.0.0
html5lib==1.1
idna==3.6
lxml==5.1.0
numpy==1.26.4
oauthlib==3.2.2
oscrypto==1.3.0
packaging==24.1
pandas==2.2.0
pillow==10.2.0
platformdirs==4.1.0
psycopg2-binary==2.9.9
pycparser==2.21
pydyf==0.9.0
pyHanko==0.23.1
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pymongo==4.6.1
pypdf==4.1.0
pyphen==0.14.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.9
requests==2.31.0
requests-oauthlib==1.3.1
setuptools==69.0.3
six==1.16.0
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.10.0
tzdata==2023.4
tzlocal==5.2
uritools==4.0.2
urllib3==2.1.0
virtualenv==20.25.0
weasyprint==61.2
webencodings==0.5.1
wheel==0.42.0
xhtml2pdf==0.2.15
zopfli==0.2.3
